+++ AoE2ScenarioParser/objects/support/area.py
@@ -274,6 +274,7 @@ class Area:
         Converts the selection to a list of OrderedSets with Tile NamedTuples with (x, y) coordinates.
         The separation between chunks is based on if they're connected to each other.
         So the tiles must share an edge (i.e. they should be non-diagonal).
+        (With exceptions like grid state when using gap_size = 0)
 
         Args:
             as_terrain: If the returning coordinates should be Tile objects or Terrain Tiles. If `True` the coordinates
@@ -302,6 +303,19 @@ class Area:
             )
 
         return chunks_ordered
+
+    def to_chunk_areas(self) -> List[Area]:
+        """
+        Converts the selection to a list of Area objects.
+        Internally, this uses `to_chunks` to convert the selection to chunks.
+        The separation between chunks is based on if they're connected to each other.
+        So the tiles must share an edge (i.e. they should be non-diagonal).
+        (With exceptions like grid state when using gap_size = 0)
+
+        Returns:
+            A list with Area objects based on the selection.
+        """
+        return [Area(corner1=chunk[0], corner2=chunk[-1]) for chunk in self.to_chunks()]
 
     def to_dict(self, prefix: str = "area_") -> Dict[str, int]:
         """